chore: refactor, introduce ClientGeneratorFactory
